숙련주차 개인과제 9단계까지 구현완료

+++ src/main/java/com/example/todolist/controller/CommentController.java
@@ -2,6 +2,7 @@ package com.example.todolist.controller;
 
 import com.example.todolist.dto.CommentRequestDto;
 import com.example.todolist.dto.CommentResponseDto;
+import com.example.todolist.exception.InvalidTokenException;
 import com.example.todolist.service.CommentService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.Authentication;
@@ -18,6 +19,9 @@ public class CommentController {
 
     @PostMapping
     public CommentResponseDto createComment(@Valid @RequestBody CommentRequestDto commentRequestDto, Authentication authentication) {
+        if (authentication == null || !authentication.isAuthenticated()) {
+            throw new InvalidTokenException("토큰이 유효하지 않습니다.");
+        }
         String username = authentication.getName();
         commentRequestDto.setUserId(username);
         return commentService.createComment(commentRequestDto);
@@ -25,13 +29,18 @@ public class CommentController {
 
     @PutMapping("/{id}")
     public CommentResponseDto updateComment(@PathVariable Long id, @Valid @RequestBody CommentRequestDto commentRequestDto, Authentication authentication) {
+        if (authentication == null || !authentication.isAuthenticated()) {
+            throw new InvalidTokenException("토큰이 유효하지 않습니다.");
+        }
         String username = authentication.getName();
-        commentRequestDto.setUserId(username);
         return commentService.updateComment(id, commentRequestDto, username);
     }
 
     @DeleteMapping("/{id}")
     public void deleteComment(@PathVariable Long id, Authentication authentication) {
+        if (authentication == null || !authentication.isAuthenticated()) {
+            throw new InvalidTokenException("토큰이 유효하지 않습니다.");
+        }
         String username = authentication.getName();
         commentService.deleteComment(id, username);
     }
